Refactor package for new SFMC Source plugin and pom changes.
