+++ src/main/java/io/cdap/plugin/sfmc/sink/DataExtensionInfo.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package io.cdap.plugin.sfmc;
+package io.cdap.plugin.sfmc.sink;
 
 import com.exacttarget.fuelsdk.ETDataExtensionColumn;
