+++ src/main/java/io/cdap/plugin/sfmc/sink/Operation.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package io.cdap.plugin.sfmc;
+package io.cdap.plugin.sfmc.sink;
 
 /**
  * The type of write operation.
